drivers: block: save return value of pci_find_capability() in u8
Callers of pci_find_capability should save the return value in u8.
change type of variables from int to u8 to match the specification.

+++ drivers/block/mtip32xx/mtip32xx.c
@@ -3936,7 +3936,7 @@ static DEFINE_HANDLER(7);
 
 static void mtip_disable_link_opts(struct driver_data *dd, struct pci_dev *pdev)
 {
-	int pos;
+	u8 pos;
 	unsigned short pcie_dev_ctrl;
 
 	pos = pci_find_capability(pdev, PCI_CAP_ID_EXP);
